refactor: #156 user repository
