+++ src/main/java/com/umc/linkyou/repository/userRepository/UserRepository.java
@@ -1,4 +1,4 @@
-package com.umc.linkyou.repository;
+package com.umc.linkyou.repository.userRepository;
 
 import com.umc.linkyou.domain.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,11 +9,8 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<Users, Long> {
+public interface UserRepository extends JpaRepository<Users, Long>,UserRepositoryCustom {
     Optional<Users> findByNickName(String nickName);
     Optional<Users> findByEmail(String email);
     Optional<Users> findById(Long id);
-    void deleteAllById(Long userId);
-
-    List<Users> findAllByStatusAndInactiveDateBefore(String inactive, LocalDateTime tenDaysAgo);
 }
