+++ src/main/java/com/umc/linkyou/service/curation/CurationServiceImpl.java
@@ -3,7 +3,6 @@ package com.umc.linkyou.service.curation;
 import com.umc.linkyou.domain.Curation;
 import com.umc.linkyou.domain.Users;
 import com.umc.linkyou.domain.classification.CurationMent;
-import com.umc.linkyou.domain.enums.CurationTopLogType;
 import com.umc.linkyou.domain.log.CurationTopLog;
 import com.umc.linkyou.repository.CurationMentRepository;
 import com.umc.linkyou.service.curation.gpt.GptService;
@@ -12,7 +11,7 @@ import com.umc.linkyou.service.curation.linku.ExternalRecommendMaterializer;
 import com.umc.linkyou.web.dto.curation.CreateCurationRequest;
 import com.umc.linkyou.repository.LogRepository.CurationTopLogRepository;
 import com.umc.linkyou.repository.CurationRepository;
-import com.umc.linkyou.repository.UserRepository;
+import com.umc.linkyou.repository.userRepository.UserRepository;
 import com.umc.linkyou.web.dto.curation.CurationDetailResponse;
 import com.umc.linkyou.web.dto.curation.CurationLatestResponse;
 import com.umc.linkyou.web.dto.curation.GptMentResponse;
@@ -21,7 +20,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.YearMonth;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
 import java.util.Random;
